refactor(ble): extract runtime modules from service files for improved modularity
Split the monolithic lifecycle_service.py, management_service.py, and receive_service.py
into separate runtime modules (lifecycle_*/management_runtime/receive_compat_service) with
a compatibility shim layer. This extracts runtime logic into focused collaborators,
improving maintainability and testability while preserving the public API surface.
