+++ meshtastic/interfaces/ble/receive_compat_service.py
@@ -1,0 +1,457 @@
+"""Receive compatibility shim service for BLE."""
+
+from typing import TYPE_CHECKING
+
+from bleak.exc import BleakError
+
+from meshtastic.interfaces.ble.client import BLEClient
+from meshtastic.interfaces.ble.receive_service import BLEReceiveRecoveryController
+
+if TYPE_CHECKING:
+    from meshtastic.interfaces.ble.coordination import ThreadCoordinator
+    from meshtastic.interfaces.ble.interface import BLEInterface
+
+class BLEReceiveRecoveryService:
+    """Service helpers for BLE receive-loop and recovery behavior."""
+
+    @staticmethod
+    def _handle_read_loop_disconnect(
+        iface: "BLEInterface", error_message: str, previous_client: BLEClient
+    ) -> bool:
+        """Handle a read-loop disconnect and return continue/stop decision.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface whose disconnect handling should be invoked.
+        error_message : str
+            Disconnect context string included in the source label.
+        previous_client : BLEClient
+            Client instance that triggered the disconnect path.
+
+        Returns
+        -------
+        bool
+            ``True`` when receive-loop processing should continue, otherwise
+            ``False``.
+        """
+        return BLEReceiveRecoveryController(iface).handle_read_loop_disconnect(
+            error_message,
+            previous_client,
+        )
+
+    @staticmethod
+    def _should_poll_without_notify(iface: "BLEInterface") -> bool:
+        """Return whether fallback polling is allowed without notify callbacks.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing notification-configuration state.
+
+        Returns
+        -------
+        bool
+            ``True`` when polling without notify callbacks should proceed,
+            otherwise ``False``.
+        """
+        with iface._state_lock:
+            return not iface._fromnum_notify_enabled
+
+    @staticmethod
+    def _coordinator_wait_for_event(
+        coordinator: "ThreadCoordinator", event_name: str, *, timeout: float | None
+    ) -> bool:
+        """Wait for a coordinator event using compatibility dispatch.
+
+        Parameters
+        ----------
+        coordinator : ThreadCoordinator
+            Coordinator instance exposing public or underscore wait helpers.
+        event_name : str
+            Event name to wait on.
+        timeout : float | None
+            Maximum wait time in seconds. ``None`` waits indefinitely when the
+            coordinator exposes ``wait_for_event``/``_wait_for_event``; when no
+            wait hook is available, fallback behavior sleeps briefly using
+            ``_sleep(COORDINATOR_WAIT_FALLBACK_SLEEP_SEC)`` and returns.
+
+        Returns
+        -------
+        bool
+            ``True`` when the event was signaled, otherwise ``False``.
+        """
+        return BLEReceiveRecoveryController._coordinator_wait_for_event(
+            coordinator,
+            event_name,
+            timeout=timeout,
+        )
+
+    @staticmethod
+    def _coordinator_check_and_clear_event(
+        coordinator: "ThreadCoordinator", event_name: str
+    ) -> bool:
+        """Check and clear a coordinator event using compatibility dispatch.
+
+        Parameters
+        ----------
+        coordinator : ThreadCoordinator
+            Coordinator instance exposing public or underscore clear helpers.
+        event_name : str
+            Event name to check and clear.
+
+        Returns
+        -------
+        bool
+            ``True`` when the event was set before clearing, otherwise
+            ``False``.
+        """
+        return BLEReceiveRecoveryController._coordinator_check_and_clear_event(
+            coordinator,
+            event_name,
+        )
+
+    @staticmethod
+    def _coordinator_clear_event(
+        coordinator: "ThreadCoordinator", event_name: str
+    ) -> None:
+        """Clear a coordinator event using compatibility dispatch.
+
+        Parameters
+        ----------
+        coordinator : ThreadCoordinator
+            Coordinator instance exposing public or underscore clear helpers.
+        event_name : str
+            Event name to clear.
+
+        Returns
+        -------
+        None
+            Returns ``None`` after best-effort event clearing.
+        """
+        BLEReceiveRecoveryController._coordinator_clear_event(
+            coordinator,
+            event_name,
+        )
+
+    @staticmethod
+    def _wait_for_read_trigger(
+        iface: "BLEInterface",
+        *,
+        coordinator: "ThreadCoordinator",
+        wait_timeout: float,
+    ) -> tuple[bool, bool]:
+        """Wait for read trigger and compute fallback poll mode.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing reconnect and notification state.
+        coordinator : ThreadCoordinator
+            Coordinator used for event waits and clear operations.
+        wait_timeout : float
+            Maximum wait time for read-trigger and reconnect events.
+
+        Returns
+        -------
+        tuple[bool, bool]
+            ``(proceed, poll_without_notify)`` where ``proceed`` indicates
+            whether the caller should continue the loop iteration.
+        """
+        return BLEReceiveRecoveryController(iface)._wait_for_read_trigger(
+            coordinator=coordinator,
+            wait_timeout=wait_timeout,
+            wait_for_event=lambda target_coordinator, event_name, timeout: BLEReceiveRecoveryService._coordinator_wait_for_event(  # noqa: E501
+                target_coordinator,
+                event_name,
+                timeout=timeout,
+            ),
+            check_and_clear_event=lambda target_coordinator, event_name: BLEReceiveRecoveryService._coordinator_check_and_clear_event(  # noqa: E501
+                target_coordinator,
+                event_name,
+            ),
+            clear_event=lambda target_coordinator, event_name: BLEReceiveRecoveryService._coordinator_clear_event(  # noqa: E501
+                target_coordinator,
+                event_name,
+            ),
+        )
+
+    @staticmethod
+    def _snapshot_client_state(
+        iface: "BLEInterface",
+    ) -> tuple[BLEClient | None, bool, bool, bool]:
+        """Snapshot client and gating flags needed by the read loop.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing state guarded by ``_state_lock``.
+
+        Returns
+        -------
+        tuple[BLEClient | None, bool, bool, bool]
+            ``(client, is_connecting, publish_pending, is_closing)``.
+        """
+        return BLEReceiveRecoveryController(iface)._snapshot_client_state()
+
+    @staticmethod
+    def _process_client_state(
+        iface: "BLEInterface",
+        *,
+        coordinator: "ThreadCoordinator",
+        wait_timeout: float,
+        client: BLEClient | None,
+        is_connecting: bool,
+        publish_pending: bool,
+        is_closing: bool,
+    ) -> bool:
+        """Process current client state and decide whether to break loop.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing reconnect and lifecycle helpers.
+        coordinator : ThreadCoordinator
+            Coordinator used for reconnect wait signaling.
+        wait_timeout : float
+            Wait timeout passed to coordinator event waits.
+        client : BLEClient | None
+            Current active BLE client snapshot.
+        is_connecting : bool
+            Whether connection establishment is currently in progress.
+        publish_pending : bool
+            Whether connect publication is pending verification.
+        is_closing : bool
+            Whether the interface is closing.
+
+        Returns
+        -------
+        bool
+            ``True`` when the caller should break out of the inner receive
+            loop, otherwise ``False``.
+        """
+        return BLEReceiveRecoveryController(iface)._process_client_state(
+            coordinator=coordinator,
+            wait_timeout=wait_timeout,
+            client=client,
+            is_connecting=is_connecting,
+            publish_pending=publish_pending,
+            is_closing=is_closing,
+            wait_for_event=lambda target_coordinator, event_name, timeout: BLEReceiveRecoveryService._coordinator_wait_for_event(  # noqa: E501
+                target_coordinator,
+                event_name,
+                timeout=timeout,
+            ),
+        )
+
+    @staticmethod
+    def _reset_recovery_after_stability(iface: "BLEInterface") -> None:
+        """Reset recovery-attempt counter after sustained stable reads.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface maintaining receive-recovery counters.
+
+        Returns
+        -------
+        None
+            Returns ``None`` after best-effort counter reset.
+        """
+        BLEReceiveRecoveryController(iface)._reset_recovery_after_stability()
+
+    @staticmethod
+    def _read_and_handle_payload(
+        iface: "BLEInterface",
+        client: BLEClient,
+        *,
+        poll_without_notify: bool,
+    ) -> bool:
+        """Read a payload iteration and dispatch packet handling.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing read/retry helpers and packet handlers.
+        client : BLEClient
+            Active BLE client used for characteristic reads.
+        poll_without_notify : bool
+            Whether the current cycle is using fallback polling mode.
+
+        Returns
+        -------
+        bool
+            ``True`` to continue the inner loop, ``False`` to stop.
+        """
+        return BLEReceiveRecoveryController(iface)._read_and_handle_payload(
+            client,
+            poll_without_notify=poll_without_notify,
+        )
+
+    @staticmethod
+    def _handle_payload_read(
+        iface: "BLEInterface",
+        client: BLEClient,
+        *,
+        poll_without_notify: bool,
+    ) -> tuple[bool, bool]:
+        """Handle one payload-read attempt and classify loop control flow.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing payload handlers and retry/disconnect helpers.
+        client : BLEClient
+            Active BLE client used for current read iteration.
+        poll_without_notify : bool
+            Whether the current cycle is running in polling fallback mode.
+
+        Returns
+        -------
+        tuple[bool, bool]
+            ``(break_inner_loop, stop_receive_loop)``.
+
+        Raises
+        ------
+        SystemExit
+            Propagated when process termination is requested.
+        KeyboardInterrupt
+            Propagated when process termination is requested.
+        """
+        return BLEReceiveRecoveryController(iface)._handle_payload_read(
+            client,
+            poll_without_notify=poll_without_notify,
+        )
+
+    @staticmethod
+    def _run_receive_cycle(
+        iface: "BLEInterface",
+        *,
+        coordinator: "ThreadCoordinator",
+        wait_timeout: float,
+    ) -> bool:
+        """Run receive-loop orchestration until shutdown or fatal stop.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface whose receive loop is being executed.
+        coordinator : ThreadCoordinator
+            Coordinator used for trigger/reconnect event waits.
+        wait_timeout : float
+            Wait timeout passed to coordinator waits for each cycle.
+
+        Returns
+        -------
+        bool
+            ``True`` when the loop exited naturally, ``False`` when a fatal
+            read path requested immediate receive-loop stop.
+        """
+        return BLEReceiveRecoveryController(iface)._run_receive_cycle(
+            coordinator=coordinator,
+            wait_timeout=wait_timeout,
+        )
+
+    @staticmethod
+    def _receive_from_radio_impl(iface: "BLEInterface") -> None:
+        """Run the BLE receive loop and dispatch decoded radio payloads.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing receive-loop state and handlers.
+
+        Returns
+        -------
+        None
+            Returns ``None``. Fatal receive-loop failures trigger recovery or
+            close side effects before exiting.
+        """
+        BLEReceiveRecoveryController(iface).receive_from_radio_impl()
+
+    @staticmethod
+    def _recover_receive_thread(iface: "BLEInterface", disconnect_reason: str) -> None:
+        """Handle receive-thread crash and guarded restart behavior.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing disconnect and receive-thread lifecycle helpers.
+        disconnect_reason : str
+            Reason label passed into disconnect handling.
+
+        Returns
+        -------
+        None
+            Returns ``None`` after scheduling recovery or stopping receive.
+        """
+        BLEReceiveRecoveryController(iface).recover_receive_thread(disconnect_reason)
+
+    @staticmethod
+    def _read_from_radio_with_retries(
+        iface: "BLEInterface",
+        client: BLEClient,
+        *,
+        retry_on_empty: bool = True,
+    ) -> bytes | None:
+        """Read non-empty payload from FROMRADIO with retry policy.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing retry policies and warning state.
+        client : BLEClient
+            Active BLE client used for characteristic reads.
+        retry_on_empty : bool
+            When ``True``, apply empty-read retry/backoff behavior.
+
+        Returns
+        -------
+        bytes | None
+            Non-empty payload bytes when available, otherwise ``None``.
+        """
+        return BLEReceiveRecoveryController(iface).read_from_radio_with_retries(
+            client,
+            retry_on_empty=retry_on_empty,
+        )
+
+    @staticmethod
+    def _handle_transient_read_error(
+        iface: "BLEInterface", error: BleakError | BLEClient.BLEError
+    ) -> None:
+        """Apply transient-read retry policy and raise on exhaustion.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface providing transient retry counters and policy helpers.
+        error : BleakError | BLEClient.BLEError
+            Error that triggered retry-policy evaluation.
+
+        Returns
+        -------
+        None
+            Returns ``None`` when retry is scheduled.
+
+        Raises
+        ------
+        BLEError
+            If transient retry attempts are exhausted.
+        """
+        BLEReceiveRecoveryController(iface).handle_transient_read_error(error)
+
+    @staticmethod
+    def _log_empty_read_warning(iface: "BLEInterface") -> None:
+        """Emit throttled warning on repeated empty FROMRADIO reads.
+
+        Parameters
+        ----------
+        iface : BLEInterface
+            Interface storing empty-read warning cooldown state.
+
+        Returns
+        -------
+        None
+            Returns ``None`` after logging or suppressing warning output.
+        """
+        BLEReceiveRecoveryController(iface).log_empty_read_warning()
